cleanup (ImageCarousel): move to `~/components/Base`
