+++ components/Base/ImageCarousel/ImageCarouselSkeleton.vue
