+++ components/Homepage/BannerCarousel/index.vue
@@ -8,7 +8,7 @@
 
 <script>
 import { mapState } from 'vuex'
-import { ImageCarousel } from '~/components/ImageCarouselV2'
+import { ImageCarousel } from '~/components/Base/ImageCarousel'
 export default {
   name: 'HomepageBannerCarousel',
   components: {
