+++ components/NewsCarousel/NewsCarousel.vue
@@ -13,7 +13,7 @@
 
 <script>
 import NewsCarouselItem from './NewsCarouselItem'
-import { ImageCarousel } from '~/components/ImageCarouselV2'
+import { ImageCarousel } from '~/components/Base/ImageCarousel'
 
 export default {
   name: 'NewsCarousel',
